Refactor: Move strategy types to CommunicationStrategy.ts
